Cleanup, first pass...
This is a terrible commit message for such a large change, but I just went through the repo and got
rid of all of the random scripts I made for testing assorted things during the development
process. I also removed a few unused files/functions + tried to reduce the amount of imported
files where possible.

A second pass (or more) is still needed before a release

Also, add mosek to dependencies since the trajectory planner will use this if clarabel fails